requirements.txt
adlfs==2022.2.0
fsspec==2023.1.0
azure-storage-blob==12.15.0
pandas==1.3.5
pytz==2021.1
pyarrow==11.0.0
kafka-python
packaging==23.0
jmespath==1.0.1
ipykernel
avro-python3
confluent-kafka

